+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # CHANGELOG
 
 - [#17] Added `get:release-line` command
+- [#16] Support createdAt and updatedAt style dates used with `dynamodb`
 
 ### 0.5.0
 
 - [#14] Add promote command
   - promotes a specific version of a built package on warehouse
   - optionally build with -b flag if not already built
-  
+
 ### 0.4.0
 
 - [#13] Add build command
@@ -26,4 +27,5 @@
 [#12]: https://github.com/warehouse.ai/wrhs/pull/12
 [#13]: https://github.com/warehouseai/wrhs/pull/13
 [#14]: https://github.com/warehouseai/wrhs/pull/14
+[#16]: https://github.com/warehouseai/wrhs/pull/16
 [#17]: https://github.com/warehouseai/wrhs/pull/17